Make the details of an XML-RPC error accessible.
This introduces an exported Error type that represents an XML-RPC error.

The calling code can try to cast the returned error to this type. If the
cast is successful, it knows that the API call happened and an XML-RPC
error was the response of the call. The Error type has getters for the
XML-RPC error code and description.